% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate_beta.R
\name{accelerate_beta}
\alias{accelerate_beta}
\title{Acceleration algorithm for betas}
\usage{
accelerate_beta(beta_list, beta, h, rhs_var)
}
\arguments{
\item{beta_list}{List of previous estimated coefficients}

\item{beta}{Vector of current estimated coefficients}

\item{h}{Acceleration iteration}

\item{rhs_var}{Righthandside variables}
}
\description{
Acceleration algorithm for betas
}
