% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coef.glmhdfe}
\alias{coef.glmhdfe}
\title{Extract coefficient estimates}
\usage{
\method{coef}{glmhdfe}(object, ...)
}
\arguments{
\item{object}{Object of class \code{glmhdfe}}

\item{...}{further arguments}
}
\description{
Extract coefficient estimates
}
