% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_vcov.R
\name{compute_vcov}
\alias{compute_vcov}
\title{Compute variance covariance matrix for standard errors from given glmhdfe_data, glmhdfe_call and glmhdfe_info objects}
\usage{
compute_vcov(data, call, info, custom_cluster = NULL,
  demean_iterations = 100, demean_tolerance = 1e-08,
  include_data_vcov = F, trace = F, verbose = F)
}
\arguments{
\item{data}{glmhdfe_data object}

\item{call}{glmhdfe_call object}

\item{info}{glmhdfe_info object}

\item{custom_cluster}{Character vector with variable names, overwrites clustering from glmhdfe_call object}

\item{demean_iterations}{Iterations for demeaning algorithm}

\item{demean_tolerance}{Demeaning tolerance}

\item{include_data_vcov}{Return data used in variance-covariance matrix estimation?}

\item{trace}{Show some information during estimation}

\item{verbose}{Show more information during estimation for the impatient}
}
\description{
Compute variance covariance matrix for standard errors from given glmhdfe_data, glmhdfe_call and glmhdfe_info objects
}
