% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demean_variable.R
\name{demean_variable}
\alias{demean_variable}
\title{Pseudo-demean variables}
\usage{
demean_variable(variable, fixed_effects, data, weights = NULL,
  demean_iterations = 100, demean_tolerance = 1e-08,
  parallel_cores = 1, suffix = "_dm", trace = F, verbose = F)
}
\arguments{
\item{variable}{Character vector of variables that should be demeaned}

\item{fixed_effects}{Character vector of fixed effects whose influence should be removed from variables}

\item{data}{Data.table with relevant data}

\item{weights}{Character string of weights for demeaning}

\item{demean_iterations}{Maximum number of demeaning iterations}

\item{demean_tolerance}{Demeaning tolerance}

\item{parallel_cores}{Number of cores to be used in demeaning. Currently fixed at 1.}

\item{suffix}{Suffix for demeaned variables.}

\item{trace}{Show some information during estimation}

\item{verbose}{Show more information during estimation for the impatient}
}
\description{
Pseudo-demean variables
}
