% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fitted.glmhdfe}
\alias{fitted.glmhdfe}
\title{Extract fitted values}
\usage{
\method{fitted}{glmhdfe}(object, ...)
}
\arguments{
\item{object}{Object of class \code{glmhdfe}}

\item{...}{further arguments}
}
\description{
Extract fitted values
}
