% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmhdfe-package.R, R/glmhdfe.R
\docType{package}
\name{glmhdfe}
\alias{glmhdfe}
\alias{glmhdfe-package}
\title{glmhdfe}
\usage{
glmhdfe(formula, data, family = gaussian(), beta = NULL,
  tolerance = 1e-08, max_iterations = 10000, fe_iterations = 5,
  accelerate = T, accelerate_iterations = 10,
  accelerate_aux_vector = T, compute_vcov = T, demean_variables = F,
  demean_iterations = 100, demean_tolerance = tolerance,
  include_fe = T, include_data = F, include_data_vcov = F,
  skip_checks = NULL, trace = F, verbose = F)
}
\arguments{
\item{formula}{Formula that describes lefthandside variable, righthandside variables of interest, sets of fixed effects and error clustering, e.g. \code{y ~ x | fe1 + fe2 | cluster1 + cluster2}}

\item{data}{\code{data.table} or \code{data.frame} with data used in the regression}

\item{family}{Estimator used, currently limited to gaussian(link = "identity"), gaussian(link = "log"), poisson(link = "log"), Gamma(link = "log") and inverse.gaussian(link = "log")}

\item{beta}{Initial beta vector, defaults to 0}

\item{tolerance}{Iteration stops when tolerance is reached (numeric)}

\item{max_iterations}{Iteration stops when maximum number of iteration is reached (integer)}

\item{fe_iterations}{Iterations for fixed effects to "catch up" when initial beta vector is provided}

\item{accelerate}{Acceleration algorithm for finding betas}

\item{accelerate_iterations}{Number of iterations before starting acceleration algorithm}

\item{accelerate_aux_vector}{Include fixed effects vectors in IRLS? Usually increases convergence speed}

\item{compute_vcov}{Compute variance-covariance matrix? Can be computed ex-post when data from estimation provided.}

\item{demean_variables}{Demean variables to be used in estimation of variance-covariance matrix?}

\item{demean_iterations}{Iterations for demeaning algorithm}

\item{demean_tolerance}{Demeaning tolerance}

\item{include_fe}{Return fixed effects?}

\item{include_data}{Return data used in estimation?}

\item{include_data_vcov}{Return data used in variance-covariance matrix estimation?}

\item{skip_checks}{Skip checks before starting procedure? Character vector can include "separation", "complete_cases", "multicollinearity".}

\item{trace}{Show some information during estimation}

\item{verbose}{Show more information during estimation for the impatient}
}
\description{
Generalized linear model with high-dimensional fixed effects

This function allows for arbitrary sets of fixed effects in a generalized linear model.
}
\author{
Julian Hinz and Joschka Wanner
}
