% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_vcov.R
\name{negative_vcov_correction}
\alias{negative_vcov_correction}
\title{Correct negative variances à la Cameron, Gelbach and Miller (2011)}
\usage{
negative_vcov_correction(x)
}
\arguments{
\item{x}{vcov matrix}
}
\description{
Correct negative variances à la Cameron, Gelbach and Miller (2011)
}
