% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{predict_convergence_time}
\alias{predict_convergence_time}
\title{Predict convergence time}
\usage{
predict_convergence_time(time, change, tolerance)
}
\arguments{
\item{time}{Vector of time, relative or absolute}

\item{change}{Vector of relative change in variable of interest}

\item{tolerance}{Tolerance for which convergence is assumed to be achieved}
}
\description{
Predict convergence time
}
