% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pretty_message}
\alias{pretty_message}
\title{Pretty messages for the impatient}
\usage{
pretty_message(verbose = F, description = NULL, time = F, task = F,
  linebreak = F, color = black, checkmark = F)
}
\arguments{
\item{verbose}{Be verbose and display messages?}

\item{description}{Message to be displayed}

\item{time}{Timestamp?}

\item{task}{Make task, i.e. dash before description}

\item{linebreak}{Break line after message?}

\item{color}{Colorful messages}

\item{checkmark}{Display checkmark at the end of message?}
}
\description{
Pretty messages for the impatient
}
