% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.glmhdfe}
\alias{print.glmhdfe}
\title{Print \code{glmhdfe} object}
\usage{
\method{print}{glmhdfe}(x, ..., digits = 4)
}
\arguments{
\item{x}{Object of class \code{glmhdfe}}

\item{...}{further arguments}

\item{digits}{Integer indicating the number of decimal places (defaults to 4)}
}
\description{
Print \code{glmhdfe} object
}
