% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.summary.glmhdfe}
\alias{print.summary.glmhdfe}
\title{Print \code{summary.glmhdfe}}
\usage{
\method{print}{summary.glmhdfe}(x, ..., digits = 4)
}
\arguments{
\item{x}{Object of class \code{summary.glmhdfe}.}

\item{...}{further arguments}

\item{digits}{Integer indicating the number of decimal places (defaults to 4)}
}
\description{
\code{print.summary.glmhdfe} is a generic function which displays summary statistics from objects
returned by \code{summary.glmhdfe}.
}
