% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{residuals.glmhdfe}
\alias{residuals.glmhdfe}
\title{Extract response residuals}
\usage{
\method{residuals}{glmhdfe}(object, ...)
}
\arguments{
\item{object}{Object of class \code{glmhdfe}}

\item{...}{further arguments}
}
\description{
Extract response residuals
}
