% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.glmhdfe}
\alias{summary.glmhdfe}
\title{Summarizing models of class \code{glmhdfe}}
\usage{
\method{summary}{glmhdfe}(object, se_type = "hessian", ...)
}
\arguments{
\item{object}{Object of class \code{glmhdfe}.}

\item{se_type}{Standard errors. Select from "hessian" (default), "robust", "outer_product_gradient", or "clustered"}

\item{...}{further arguments}
}
\value{
Returns an object of class \code{summary.glmhdfe} which is a list of summary statistics of
\code{object}.
}
\description{
Summary statistics for objects of class \code{glmhdfe}.
}
